Change data model
It was unnecessary to store the whole representation of the
feed items, as we load those from the RSS feed itself as the
user clicks on it.

Also renamed the `items` table to `readItems`, as it now only
stores the id and parent id of items that were already read
so that we can filter those out when loading the RSS feed.
To mark an item as `unread`, simply delete that item from the
database.

Now we always load the feed from its originating URL.